# Move field paths cache out of global state