@pelletier 2021-05-07
on my desktop:

```
name                              old time/op    new time/op    delta
UnmarshalDataset/config-32          88.1ms ± 2%    87.0ms ± 3%     ~     (p=0.123 n=10+10)
UnmarshalDataset/canada-32           112ms ± 3%     111ms ± 4%     ~     (p=0.579 n=10+10)
UnmarshalDataset/citm_catalog-32    60.5ms ± 3%    58.7ms ± 5%   -2.88%  (p=0.017 n=9+10)
UnmarshalDataset/twitter-32         28.3ms ± 3%    26.5ms ± 2%   -6.13%  (p=0.000 n=9+8)
UnmarshalDataset/code-32             335ms ± 6%     321ms ± 7%   -4.19%  (p=0.029 n=10+10)
UnmarshalDataset/example-32          540µs ± 4%     490µs ±13%   -9.32%  (p=0.000 n=10+10)
UnmarshalSimple-32                  1.45µs ± 5%    2.09µs ±13%  +44.79%  (p=0.000 n=10+10)
ReferenceFile-32                    47.1µs ± 4%    60.0µs ± 8%  +27.31%  (p=0.000 n=10+10)

name                              old speed      new speed      delta
UnmarshalDataset/config-32        11.9MB/s ± 2%  12.1MB/s ± 3%     ~     (p=0.101 n=10+10)
UnmarshalDataset/canada-32        19.7MB/s ± 3%  19.8MB/s ± 5%     ~     (p=0.579 n=10+10)
UnmarshalDataset/citm_catalog-32  9.23MB/s ± 2%  9.51MB/s ± 5%   +3.00%  (p=0.022 n=9+10)
UnmarshalDataset/twitter-32       15.6MB/s ± 3%  16.6MB/s ± 2%   +6.52%  (p=0.000 n=9+8)
UnmarshalDataset/code-32          8.01MB/s ± 6%  8.37MB/s ± 7%   +4.43%  (p=0.027 n=10+10)
UnmarshalDataset/example-32       15.0MB/s ± 4%  16.6MB/s ±14%  +10.71%  (p=0.000 n=10+10)
ReferenceFile-32                   111MB/s ± 4%    88MB/s ± 9%  -21.35%  (p=0.000 n=10+10)

name                              old alloc/op   new alloc/op   delta
UnmarshalDataset/config-32          16.9MB ± 0%    16.9MB ± 0%   +0.00%  (p=0.015 n=10+10)
UnmarshalDataset/canada-32          74.3MB ± 0%    74.3MB ± 0%   +0.00%  (p=0.000 n=10+10)
UnmarshalDataset/citm_catalog-32    37.1MB ± 0%    37.1MB ± 0%   +0.00%  (p=0.009 n=10+10)
UnmarshalDataset/twitter-32         15.6MB ± 0%    15.6MB ± 0%     ~     (p=1.000 n=9+10)
UnmarshalDataset/code-32            59.3MB ± 0%    59.3MB ± 0%   +0.00%  (p=0.001 n=9+10)
UnmarshalDataset/example-32          238kB ± 0%     238kB ± 0%   +0.06%  (p=0.000 n=9+10)
ReferenceFile-32                    11.8kB ± 0%    20.8kB ± 0%  +76.07%  (p=0.000 n=10+10)

name                              old allocs/op  new allocs/op  delta
UnmarshalDataset/config-32            645k ± 0%      645k ± 0%   +0.00%  (p=0.012 n=10+10)
UnmarshalDataset/canada-32            896k ± 0%      896k ± 0%   +0.00%  (p=0.000 n=10+10)
UnmarshalDataset/citm_catalog-32      370k ± 0%      370k ± 0%   +0.00%  (p=0.001 n=10+10)
UnmarshalDataset/twitter-32           157k ± 0%      157k ± 0%     ~     (p=0.710 n=9+10)
UnmarshalDataset/code-32             2.91M ± 0%     2.91M ± 0%   +0.00%  (p=0.032 n=10+10)
UnmarshalDataset/example-32          3.63k ± 0%     3.63k ± 0%   +0.03%  (p=0.000 n=10+10)
ReferenceFile-32                       253 ± 0%       387 ± 0%  +52.96%  (p=0.000 n=10+10)
```

I'm expecting this to be a pretty significant performance drop, as a lot of benchmarks use structs. The reason why this is a global state is that I've found in practice to trade off a bit of converging memory amount is a good tradeoff for the performance it gives. Being global ensures that the convergence is reached faster (in terms of calls to Decode), and allows Unmarshal to leverage this cache.

What's the motivation behind removing it? If it's about the start time impact maybe we can find a way to only put the bare minimum in the global state (just 32bits?), and only allocate the map when first used?