@moorereason 2021-05-12
My only motivation was to remove global state.  If the cache is a significant performance gain, then it's worth keeping.