@pelletier 2021-05-14
This cache results on significant performance gains for any program that needs to unmarshal into the same struct type more than once (which sounds like a lot of programs?). I think it's worth keeping it, as the alternative would be to create an API for the user to attach its own cache manually, which seems like a niche concern and a big commitment to backward compatibility.